requirements.txt
MetaTrader5>=5.0.37,<6.0
pandas
numpy
ta
river>=0.12.0,<0.13.0
python-telegram-bot>=20.0,<21.0

PyYAML
jsonschema
matplotlib
psutil
schedule

pytest
pytest-mock
coverage

# Optional for Linting & Typing
flake8
mypy 